refactor: make i18n pkg export everything
To alleviate the need for including full paths in imports.
