+++ packages/portal/src/cachers/utils.js
@@ -2,7 +2,7 @@ import axios from 'axios';
 import { createClient } from 'redis';
 import _pick from 'lodash/pick.js';
 import { daily } from '../plugins/europeana/utils.js';
-import { langMapValueForLocale } from '@europeana/i18n/src/langMap.js';
+import { langMapValueForLocale } from '@europeana/i18n';
 import EuropeanaRecordApi from '../plugins/europeana/record.js';
 import EuropeanaEntityApi from '../plugins/europeana/entity.js';
