+++ packages/portal/src/components/embed/EmbedMap.vue
@@ -29,7 +29,7 @@
 
 <script>
   import langAttributeMixin from '@/mixins/langAttribute';
-  import { langMapValueForLocale } from '@europeana/i18n/src/langMap.js';
+  import { langMapValueForLocale } from '@europeana/i18n';
 
   export default {
     name: 'EmbedMap',
