+++ packages/portal/src/components/item/ItemPreviewCard.vue
@@ -70,7 +70,7 @@
 </template>
 
 <script>
-  import { langMapValueForLocale } from '@europeana/i18n/src/langMap.js';
+  import { langMapValueForLocale } from '@europeana/i18n';
 
   import ContentCard from '../content/ContentCard';
