+++ packages/portal/src/pages/contentful/record-harvester/index.vue
@@ -33,7 +33,7 @@
 </template>
 
 <script>
-  import { langMapValueForLocale } from '@europeana/i18n/src/langMap.js';
+  import { langMapValueForLocale } from '@europeana/i18n';
   import { recordIdFromUrl } from '@/plugins/europeana/record';
 
   export default {
